pyproject.toml
[project]
name = "langchain-playground"
version = "0.1.0"
description = "Test, study, and experimentation with LangChain and LangGraph."
authors = [
    {name = "miche1696",email = "makalele96@gmail.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph (>=0.5.2,<0.6.0)",
    "openai (>=1.95.1,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pytube (>=15.0.0,<16.0.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)"
]

[tool.poetry]
packages = [{include = "langchain_playground", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.4.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
